pyproject.toml
[project]
name = "csp-vms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-mgmt-compute>=37.0.1",
    "azure-mgmt-costmanagement>=4.0.1",
    "azure-mgmt-network>=29.0.0",
    "azure-mgmt-resource>=24.0.0",
    "boto3>=1.40.55",
    "google-cloud-compute>=1.40.0",
    "google-cloud-resource-manager>=1.14.2",
    "openpyxl>=3.1.5",
    "tabulate>=0.9.0",
]
